+++ _components/01.rotating-elements/js/rotating-elements.js
@@ -28,7 +28,7 @@ App = ( function ( App, $, window, document ) {
 						mouseX = pointerEvent.pageX;
 						mouseY = pointerEvent.pageY;
 					} else {
-						mouseX = e.clientX,
+						mouseX = e.clientX;
 						mouseY = e.clientY;
 					}