+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 #### = 1.5.3 (May 10, 2018) =
 
-
+* New: Add a new module "3D Background".
 * New: Add a new module "Dotted Line".
 * New: Add a new module "Rotating Elements".
 * New: Add a new module "Ajax Page Loader".